refactor: move pkg/{agent,mongot,tls} from MCO to root (Plan 4 of MCO decoupling)
Relocates three shared packages from mongodb-community-operator/pkg/ to root pkg/:

- pkg/agent: readiness/health polling helpers + agent flags; moved wholesale,
  no code changes. 1 Enterprise importer + 2 MCO-internal importers updated.
- pkg/mongot: mongot YAML config types; moved wholesale, no code changes.
  3 Enterprise importers updated.
- pkg/tls (MCO): TLS secret management (EnsureTLSSecret, TLSConfigurableResource,
  OperatorSecretFileName). The root pkg/tls already exists (Enterprise TLS mode
  helpers); no symbol conflicts exist, so the MCO content was added as
  pkg/tls/tls_secret.go within the same package. 1 Enterprise importer updated.

No transitive MCO subpackage deps discovered (all three packages are pure leaves).
Path-normalised test diff vs Plan 3 baseline is empty (74 ok / 0 FAIL).

Spec: docs/design/decouple-mdb-community-investigation.md clusters 9, 10, 11.
Stacked on Plan 3 (PR #1090, branch decuple-plan-3-api-common).
